GROK-14366: EDA | Missing Vals Imput: KNN fails imputation
Values, failed to be imputed by KNN, are imputed
