+++ packages/EDA/src/missing-values-imputation/ui.ts
@@ -2,22 +2,15 @@ import * as grok from 'datagrok-api/grok';
 import * as ui from 'datagrok-api/ui';
 import * as DG from 'datagrok-api/dg';
 
-import { TITLE, LINK, ERROR_MSG, HINT, CONTENT } from './ui-constants';
+import { TITLE, LINK, ERROR_MSG, HINT } from './ui-constants';
 import { SUPPORTED_COLUMN_TYPES, METRIC_TYPE, DISTANCE_TYPE, MetricInfo, DEFAULT, MIN_NEIGHBORS, 
-  impute, getMissingValsIndices, areThereFails } from "./knn-imputer";
-import { FILL_VALUE, SimpleImputTask, simpleImput } from "./simple-imputer";
+  impute, getMissingValsIndices, areThereFails, imputeFailed } from "./knn-imputer";
 
 /** Setting of the feature metric inputs */
 type FeatureInputSettings = {
   defaultWeight: number,
   defaultMetric: METRIC_TYPE,
   availableMetrics: METRIC_TYPE[], 
-};
-
-/** Settings for simple imputation */
-type SimpleImputSettings = {
-  default: FILL_VALUE,
-  available: FILL_VALUE[]
 };
 
 /** Return default setting of the feature metric inputs */
@@ -38,29 +31,6 @@ function getFeatureInputSettings(type: DG.COLUMN_TYPE): FeatureInputSettings {
         defaultWeight: DEFAULT.WEIGHT,
         defaultMetric: METRIC_TYPE.DIFFERENCE,
         availableMetrics: [METRIC_TYPE.DIFFERENCE, METRIC_TYPE.ONE_HOT]
-      };
-
-    default:
-      throw new Error(ERROR_MSG.UNSUPPORTED_COLUMN_TYPE);
-  }
-}
-
-/** */
-function getSimpleMethodSettings(type: DG.COLUMN_TYPE): SimpleImputSettings {
-  switch (type) {
-    case DG.COLUMN_TYPE.STRING:
-    case DG.COLUMN_TYPE.DATE_TIME:
-      return {
-        default: FILL_VALUE.MOST_FREQUENT,
-        available: [FILL_VALUE.MOST_FREQUENT]
-      };
-
-    case DG.COLUMN_TYPE.INT:
-    case DG.COLUMN_TYPE.FLOAT:
-    case DG.COLUMN_TYPE.QNUM:
-      return {
-        default: FILL_VALUE.MEAN,
-        available: [FILL_VALUE.MOST_FREQUENT, FILL_VALUE.MEAN, FILL_VALUE.MEDIAN]
       };
 
     default:
@@ -253,7 +223,7 @@ export function runKNNImputer(): void {
         const failedToImpute = impute(df!, targetColNames, featuresMetrics, distType, neighbors, inPlace);
 
         if (!keepEmpty)
-          processFailedImputations(df!, failedToImpute);
+          imputeFailed(df!, failedToImpute);
       }
       catch (err) {
         if (err instanceof Error)
@@ -271,116 +241,3 @@ export function runKNNImputer(): void {
     .add(keepEmptyInput)
     .show();
 } // runKNNImputer
-
-/** Run the simple missing values imputer*/
-function simpleImputation(df: DG.DataFrame, colsWithMissingVals: Map<string, number[]>): void {
-  const len = colsWithMissingVals.size;
-
-  if (len > 0) {        
-    const selected = new Map<string, boolean>();
-    const imputType = new Map<string, FILL_VALUE>();
-    const imputTypeElems = new Map<string, HTMLElement>();
-
-    const divs = [] as any[];
-
-    // create UI for imputation specifying
-    colsWithMissingVals.forEach((ignore, name) => {
-      selected.set(name, DEFAULT.SELECTED > 0);
-
-      const settings = getSimpleMethodSettings(df.col(name)!.type as DG.COLUMN_TYPE);
-      imputType.set(name, settings.default);
-
-      // UI for type of imputation
-      const imputTypeInput = ui.choiceInput('', settings.default, settings.available, () => {
-        imputType.set(name, imputTypeInput.value ?? settings.default);        
-      });
-      imputTypeInput.root.hidden = true;
-      imputTypeElems.set(name, imputTypeInput.root);
-      imputTypeInput.setTooltip(HINT.FILL_VALUE);
-      
-      // UI for column selection
-      const switcher = ui.switchInput(name, DEFAULT.SELECTED > 0, () => {
-        const val = switcher.value;
-        selected.set(name, val);
-        imputTypeInput.root.hidden = !(val && toShowImputTypeElems);
-        checkSelected();
-      });
-      switcher.setTooltip(`nulls: ${df.col(name)!.stats.missingValueCount}`);
-
-      divs.push(ui.divH([switcher.root, imputTypeInput.root]));
-    });
-
-    /** disable OK button if nothing is selected */
-    const checkSelected = () => {
-      let toDisable = false;
-      selected.forEach((val) => toDisable = toDisable || val);
-      dlg.getButton(TITLE.OK).disabled = !toDisable;
-    };
-
-    let toShowImputTypeElems = false;
-
-    /* settings icon: show/hide inputs for imputation type */
-    const settingsIcon = ui.icons.settings(() => {
-      toShowImputTypeElems = !toShowImputTypeElems;
-
-      colsWithMissingVals.forEach((val, name) => {
-        imputTypeElems.get(name)!.hidden = !(toShowImputTypeElems && selected.get(name)!);
-      });
-    }, HINT.IMPUTATION_SETTINGS);
-    
-
-    const dlg = ui.dialog({title: TITLE.SIMPLE_IMPUTER, helpUrl: LINK.SIMPLE_IMPUTER});
-    grok.shell.v.root.appendChild(dlg.root);
-
-    dlg.onOK(() => {
-      dlg.close();
-
-      // task for simple imputer
-      const taskMap = new Map<string, SimpleImputTask>();
-
-      // create task
-      colsWithMissingVals.forEach((indeces, name) => {
-        if (selected.get(name)) {
-          taskMap.set(name, {indeces: indeces, fillValueType: imputType.get(name)!})
-      }});
-
-      // perform imputation
-      try {
-        simpleImput(df, taskMap);
-      }
-      catch (err) {
-        if (err instanceof Error)
-          grok.shell.error(`${ERROR_MSG.KNN_FAILS}: ${err.message}`);
-        else
-          grok.shell.error(`${ERROR_MSG.KNN_FAILS}: ${ERROR_MSG.CORE_ISSUE}`);
-      } 
-
-    });
-    
-    dlg.add(ui.divH([ui.divText(CONTENT.FILL_MISS_VALS), settingsIcon]))
-      .add(ui.divV(divs))
-      .show();
-
-    checkSelected();
-  }
-} // simpleImputation
-
-/** Process items that are failed to be imputed */
-function processFailedImputations(df: DG.DataFrame, failedToImpute: Map<string, number[]>) {
-  const len = failedToImpute.size;
-
-  if (len > 0) {
-    /*const fillBtn = ui.bigButton(TITLE.FILL, () => { simpleImputation(df, failedToImpute); }, HINT.FILL_FAILED_ITEMS);
-
-    const markBtn = ui.bigButton(TITLE.MARK, () => {
-      alert('Hi');
-    }, HINT.MARK_FAILED_ITEMS);
-
-    grok.shell.warning(ui.divV([
-      `${ERROR_MSG.FAILED_TO_IMPUTE} ${len} values`,
-      ui.divH([markBtn, fillBtn]),
-    ]));*/
-
-    simpleImputation(df, failedToImpute);
-  }
-}
